Move utility method into the tests folder
