+++ common/src/main/java/org/hiero/mirror/common/util/DomainUtils.java
@@ -218,10 +218,6 @@ public class DomainUtils {
         return convertToNanosMax(timestamp.getSeconds(), timestamp.getNanos());
     }
 
-    public static long getUnixSeconds(long nanoseconds) {
-        return nanoseconds / NANOS_PER_SECOND;
-    }
-
     /**
      * Cleans a string of invalid characters that would cause it to fail when inserted into the database. In particular,
      * PostgreSQL does not allow the null character (0x0000) to be inserted.
